requirements.txt
django==5.0.6
django-ninja==1.1.0
jwt==1.3.1